Created Rudimentary Dash State Wall Jump + Others
Wall Jump for Dash State Crudely Implemented, plays poorly with current input consumption

Created Dampen Air Accel Feature of The MT Character Controller, calling this method sets it to a low excel mode for x seconds, would be better if there was a lerping affect with good ux parameters via calculations.

Perhaps the Movementstate could have a get air acceleration that the player uses, but we'll have to wait until a major organization refactor happens before we mess with that.